refactor: refactor landing page for server component
