+++ apps/web/app/hooks/use-toast.ts
@@ -1,3 +1,4 @@
+'use client'
 // Inspired by react-hot-toast library
 import * as React from 'react'
