+++ apps/web/app/page.tsx
@@ -1,26 +1,23 @@
-'use client'
-import { QueryClient, QueryClientProvider } from 'react-query'
+import ScrollToTopButton from './components/ui/ScrollToTopButton'
+import { FAQ } from './landing-page/FAQ'
 import { Header } from './landing-page/Header'
 import { Hero } from './landing-page/Hero'
-import { Values } from './landing-page/Values'
-import { FAQ } from './landing-page/FAQ'
 import { HowItWorks } from './landing-page/HowItWorks'
-import ScrollToTopButton from './components/ui/ScrollToTopButton'
+import { Values } from './landing-page/Values'
 
-export default function Page() {
-  const queryClient = new QueryClient()
+export const revalidate = 5
+
+export default async function Page() {
   return (
     <>
-      <QueryClientProvider client={queryClient}>
-        <Header />
-        <main>
-          <Hero />
-          <Values />
-          <HowItWorks />
-          <FAQ />
-          <ScrollToTopButton />
-        </main>
-      </QueryClientProvider>
+      <Header />
+      <main>
+        <Hero />
+        <Values />
+        <HowItWorks />
+        <FAQ />
+        <ScrollToTopButton />
+      </main>
     </>
   )
 }
